DT-33: Rename TVPMain.Forms namespace
The TVPMain.Form namespace was conflicting with the
System.Windows.Forms namespace, which caused every
new form to fail until the "Form" type had been properly
namespaced.

Since this conflicts with default Visual Studio behavior,
and the namespace was more appropriate as a plural,
it seemed best to rename the namespace.